rename analyse to analyze
