+++ src/interpreter.ts
@@ -4,7 +4,7 @@
 import prompts from 'prompts';
 import { Token } from './types/tokenTypes';
 import { emptyEnvironment } from './modules/value';
-import lexicalAnalyse from './modules/lexical-analyse';
+import lexicalAnalyse from './modules/lexical-analyze';
 import parseSource from './modules/statementParser';
 import evaluate from './modules/evaluator';
